% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_functions.R
\name{find_other_studies_evaluating_same_germplasm}
\alias{find_other_studies_evaluating_same_germplasm}
\title{For a study_id, find other studies having a minimum number common accessions}
\usage{
find_other_studies_evaluating_same_germplasm(
  study_id,
  brapi_conn,
  min_germ_common = 5
)
}
\arguments{
\item{study_id}{A BrAPI study IDs to query}

\item{brapi_conn}{A BrAPI connection object, created by
\code{BrAPI::createBrAPIConnection()}}

\item{min_germ_common}{An integer of the minimum number of germplasms in
common for the study to be included in the response}
}
\value{
A tibble with one row per trial
other_studies: the study_db_id and
n: the number of germplasms in common with the focal study
}
\description{
For a BrAPI study ID, get the germplasm evaluated in it
Find other studies where those same germplasms were evaluated
Select studies where a minimum number of germplasms are in common with the
focal study ID
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("T3/Wheat", is_breedbase = TRUE)

# Retrieve other studies based on a focal study
df <- find_other_studies_evaluating_same_germplasm("10674", brapi_conn)
df
}

}
