% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{get_geno_protocol_from_single_germ}
\alias{get_geno_protocol_from_single_germ}
\title{Get genotyping protocol metadata for a single germplasm}
\usage{
get_geno_protocol_from_single_germ(germ_id, brapi_connection, verbose = F)
}
\arguments{
\item{germ_id}{The germplasmDbId for the accession of interest.}

\item{brapi_connection}{A BrAPI connection object, typically from
\code{BrAPI::createBrAPIConnection()}, with a \code{$wizard()} method.}

\item{verbose}{Logical; if TRUE, prints progress messages.}
}
\value{
A tibble with a single row containing
\code{germplasm_db_id}, \code{genoProtocol_db_id}, and \code{genoProtocolName}.
The genotyping protocol columns are list columns, potentially containing
multiple protocol IDs/names.
}
\description{
Queries the T3 AJAX interface using the \code{$wizard()} method of a
brapi_connection to determine which genotyping protocols have been used for a
specific germplasm
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

winner_geno_protocols <- get_geno_protocol_from_germ_vec("1284387", brapi_conn)
winner_geno_protocols
}

}
