% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{get_germplasm_from_trial_vec}
\alias{get_germplasm_from_trial_vec}
\title{Get germplasm metadata for multiple trials}
\usage{
get_germplasm_from_trial_vec(study_id_vec, brapi_connection, verbose = F)
}
\arguments{
\item{study_id_vec}{A character vector of studyDbIds to query.}

\item{brapi_connection}{A BrAPI connection object as used in
\code{get_germplasm_from_single_trial()}.}

\item{verbose}{Logical; passed on to \code{get_germplasm_from_single_trial()} to
control logging. If FALSE display purrr progress bar}
}
\value{
A data frame obtained by row-binding the results of each trial, with
one row per germplasm per trial
}
\description{
Wrapper around \code{\link{get_germplasm_from_single_trial}} to retrieve and combine
germplasm metadata for a vector of trial IDs.
}
\examples{
\dontrun{
brapi_conn <- BrAPI::getBrAPIConnection("T3/Wheat")

all_germ <- T3BrapiHelpers::get_germplasm_from_trial_vec(
  c("10673", "10677"), brapi_conn)
all_germ
}

}
