% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_trials.R
\name{get_lat_long_elev_from_location_vec}
\alias{get_lat_long_elev_from_location_vec}
\title{Get location info from a vector of locations via BrAPI}
\usage{
get_lat_long_elev_from_location_vec(
  loc_vec,
  brapi_connection,
  id_or_name = "name"
)
}
\arguments{
\item{loc_vec}{A vector of location names or DB IDs for which you want lat,
long, and elevation values}

\item{brapi_connection}{A BrAPI connection object, typically from
\code{BrAPI::createBrAPIConnection()}, with a \code{$search()} method.}

\item{id_or_name}{A string. If "name" will expect loc_vec to be a vector of
location names else a vector of location DB IDs.}
}
\value{
A data frame of lattitude, longitude and elevation values for the
loactions
}
\description{
Queries the BrAPI \code{/search/locations} endpoint and
returns a data frame of lat, long, and elevation values for those locations
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("wheat.triticeaetoolbox.org", is_breedbase = TRUE)

locs_df <- get_lat_long_elev_from_location_vec(c("31", "143"), brapi_conn)
locs_df
}

}
