% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_trials.R
\name{get_traits_from_single_trial}
\alias{get_traits_from_single_trial}
\title{Get traits measured from a single trial via BrAPI}
\usage{
get_traits_from_single_trial(study_id, brapi_connection, verbose = F)
}
\arguments{
\item{study_id}{A single studyDbId to query germplasm for.}

\item{brapi_connection}{A BrAPI connection object, typically from
\code{BrAPI::createBrAPIConnection()}, with a \code{$search()} method.}

\item{verbose}{Logical; if \code{TRUE}, print messages about the retrieval
process.}
}
\value{
A data frame of traits for the given trial, with one row per trait
Columns include \code{observationVariable_db_id} and
\code{observation_variable_name}. If no result is found, not sure what
happens...
}
\description{
Queries the BrAPI \code{/search/variables} endpoint for a given trial and
returns a data frame of traits and their DbIds measured in that trial
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

traits_df <- get_traits_from_single_trial("8128", brapi_conn)
traits_df
}

}
