% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_trials.R
\name{get_traits_from_trial_vec}
\alias{get_traits_from_trial_vec}
\title{Retrieve what traits were measured for a set of trials by study IDs}
\usage{
get_traits_from_trial_vec(
  study_id_vec,
  brapi_connection,
  id_or_name = "name",
  verbose = F
)
}
\arguments{
\item{study_id_vec}{A character vector of BrAPI study IDs (studyDbId values)
to query.}

\item{brapi_connection}{A BrAPI connection object, typically created by
\code{BrAPI::createBrAPIConnection()}, with \code{$search()} method.}

\item{id_or_name}{A string. If "name" return the names of the traits else
return the trait DB IDs.}

\item{verbose}{A logical. If TRUE a lot of info on the traits in the studies
else a purrr progress bar}
}
\value{
A vector of either trait names or trait DB IDs.
}
\description{
Given a vector of BrAPI study IDs, use the search function of a Breedbase
connection to compile a vector of all traits measured in each trial in the
study_id_vec
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

traits <- get_traits_from_trial_vec(c("8128", "9421"), brapi_conn)
traits
}

}
