% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_germplasm.R
\name{get_trial_from_germplasm_vec}
\alias{get_trial_from_germplasm_vec}
\title{Get trial metadata for multiple germplasms}
\usage{
get_trial_from_germplasm_vec(germplasm_id_vec, brapi_connection, verbose = F)
}
\arguments{
\item{germplasm_id_vec}{A character vector of germplasmDbIds to query.}

\item{brapi_connection}{A BrAPI connection object as used in
\code{get_trial_from_single_germplasm()}.}

\item{verbose}{Logical; passed on to \code{get_trial_from_single_germplasm()} to
control logging. If FALSE display purrr progress bar}
}
\value{
A data frame obtained by row-binding the results of each germplasm,
with one row per trial per germplasm
}
\description{
Wrapper around \code{\link{get_trial_from_single_germplasm}} to retrieve and
combine trial metadata for a vector of germplasm IDs.
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("wheat.triticeaetoolbox.org", is_breedbase = TRUE)

all_trial <- get_trial_from_germplasm_vec(c("1284387", "1382716"), brapi_conn)
all_trial
}

}
