% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_trials.R
\name{get_trial_meta_data_from_trial_vec}
\alias{get_trial_meta_data_from_trial_vec}
\title{Retrieve metadata for a set of trials by study IDs}
\usage{
get_trial_meta_data_from_trial_vec(study_id_vec, brapi_connection)
}
\arguments{
\item{study_id_vec}{A character vector of BrAPI study IDs (studyDbId values)
to query.}

\item{brapi_connection}{A BrAPI connection object, typically created by
\code{BrAPI::createBrAPIConnection()},
with \code{$get()} method available.}
}
\value{
A tibble-like data frame with one row per trial and cleaned column
names (via \code{janitor::clean_names()}). Date columns \code{start_date}
and \code{end_date} are converted to \code{POSIXct} in UTC.
}
\description{
Given a vector of BrAPI study IDs, query the \code{/studies/{studyDbId}}
endpoint for each and compile a tidy data frame of trial metadata.
}
\examples{
\dontrun{
brapi_conn <- BrAPI::createBrAPIConnection("wheat-sandbox.triticeaetoolbox.org", is_breedbase = TRUE)

# Retrieve metadata for two trials
df <- get_trial_meta_data_from_trial_vec(c("8128", "9421"), brapi_conn)
df
}

}
